+++ slipp/src/test/java/slipp/dao/JdbcUserDaoTest.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class UserDaoTest {
+public class JdbcUserDaoTest {
     @BeforeEach
     public void setup() {
         ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
@@ -24,7 +24,7 @@ public class UserDaoTest {
     @Test
     public void crud() throws Exception {
         User expected = new User("userId", "password", "name", "javajigi@email.com");
-        UserDao userDao = new UserDao();
+        UserDao userDao = new JdbcUserDao();
         userDao.insert(expected);
         User actual = userDao.findByUserId(expected.getUserId()).get();
         assertThat(actual).isEqualTo(expected);
@@ -37,8 +37,8 @@ public class UserDaoTest {
 
     @Test
     public void findAll() throws Exception {
-        UserDao userDao = new UserDao();
-        List<User> users = userDao.findAll();
+        JdbcUserDao jdbcUserDao = new JdbcUserDao();
+        List<User> users = jdbcUserDao.findAll().get();
         assertThat(users).hasSize(1);
     }
 }
